Accident-Detector/requirements.txt
# Core Dependencies
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
matplotlib>=3.7.0
requests>=2.31.0
twilio>=8.10.0
python-dotenv>=1.0.0

# Streamlit and Web Framework
streamlit>=1.31.0
streamlit-folium>=0.16.0

# Map and Location Services
folium>=0.15.0

# Additional utilities
pandas>=2.0.0
scipy>=1.10.0